+++ Classes/Service/MatchService.php
@@ -8,7 +8,7 @@ use Sys25\RnBase\Search\SearchBase;
 use Sys25\RnBase\Typo3Wrapper\Service\AbstractService;
 use System25\T3sports\Model\Competition;
 use System25\T3sports\Model\CompetitionRound;
-use System25\T3sports\Model\Match;
+use System25\T3sports\Model\Fixture;
 use System25\T3sports\Model\MatchNote;
 use System25\T3sports\Model\Repository\MatchNoteRepository;
 use System25\T3sports\Model\Repository\MatchRepository;
@@ -92,7 +92,7 @@ class MatchService extends AbstractService
      * @param string $teamIds
      * @param string $status
      *
-     * @return Match[]
+     * @return Fixture[]
      */
     public function getMatches4Competition($comp, $teamIds = '', $status = '0,1,2')
     {
@@ -228,7 +228,7 @@ class MatchService extends AbstractService
      * Wenn $types = 1 dann
      * werden nur die Notes mit dem Typ != 100 geliefert.
      *
-     * @param Match $match
+     * @param Fixture $match
      * @param bool $excludeTicker
      *
      * @return MatchNote[]
@@ -263,6 +263,6 @@ class MatchService extends AbstractService
      */
     public function getMatchTable()
     {
-        return \tx_rnbase::makeInstance(MatchTableBuilder::class);
+        return tx_rnbase::makeInstance(MatchTableBuilder::class);
     }
 }